Add imx-bootlet-header recipe
In order to but the MX28 with HAD enabled you need
a small header before the bootstream.  We copied the
FSL provided mk_hdr.sh script into the build as a native
package so it's available to the image class.  This allows
us to generate a header for the SD card.

Fix the reach image class to use the mxshdr.sh script and
adjust the dd commands for the SD card image

This is a working baseline for Canby with one working image
reach-image-minimal.

Signed-off-by: Jeff Horn <jhorn@reachtech.com>