udev: fix broken dev cache, enable dev cache
The init script support for dev caching was broken, resulting in
*longer* boot times if caching was enabled.

This patch fixes the caching support.

The faster startup time sometimes tickles a race in
newly introduced code to support the transition from
/dev/.udev to /run/udev for the runtime directory.

This results in the somewhat scary error message:

udevd[591]: error: runtime directory '/run/udev' not writable, for now falling back to '/dev/.udev'

This is not actually an error, just a notification that udev
is using its historical /dev/.udev directory rather than
the new /run/udev.

This patch changes the message from an error call to an info
call.

Signed-off-by: Steve Sakoman <steve@sakoman.com>