busybox: Convert recipe changes to configuration fragments
The configuration fragments mode of adding overrides on the original
set of features of busybox is more scalable and straightforward across
releases. This allow us to change the bbappend to be version agnostic.

The enabled features are:

 - resize - from O.S. Systems Embedded Linux
 - rdev
 - ifplugd
 - mdev

The mdev.conf file override is dropped as the OE-Core one provides the
touchscreen support and also a set of default settings thus providing
better Linux userspace compatibility.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>