Add a check for the 10.4" Sciton machines
We currently hack our ts.conf file into place in pointercal's init
script. It was making an i2c call to see if we were using pcap, and
defaulting to resistive if not. This adds a usb call to check for dewar
USB touch controllers, and if present, use the correct pcap ts.conf.

Signed-off-by: Stephano Cetola <scetola@reachtech.com>