reach-installer: Add recipe
reach-installer starts at boot and runs reach-installer script to
format and copy files from /installer-data to nand and nor flash.
U-boot is copy to nor and kernel, dtb and rootfs are copy to nand.
The installer-data directory is created on sd card to store files
to be copy to mtd device.

After transfer is completed the systems is power off. To boot,
close pins 2-3 of jumper JP1 and power up the board.

Default gh2 board layout:

    /dev/mtd0 - dtb partion     (NAND)
    /dev/mtd1 - kernel partion  (NAND)
    /dev/mtd2 - rootfs partion  (NAND)
    /dev/mtd3 - boot partion    (NOR)
    /dev/mtd4 - bootenv partion (NOR)

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>