Revert back to tslib for touchscreen support
The current touchscreen devices are single-touch and requires
calibration to proper work, for such we need to revamp the system and
move to libinput however due deadlines this will be left for future.

This change groups all the tslib related changes so in future we can
easily revert it again.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>